+++ koin-projects/examples/androidx-compose-jetnews/src/main/java/com/example/jetnews/ui/home/PostCards.kt
@@ -16,25 +16,25 @@
 
 package com.example.jetnews.ui.home
 
-import androidx.compose.foundation.Icon
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.Text
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.preferredSize
-import androidx.compose.material.AmbientEmphasisLevels
+import androidx.compose.material.AmbientContentAlpha
+import androidx.compose.material.ContentAlpha
+import androidx.compose.material.Icon
 import androidx.compose.material.IconToggleButton
 import androidx.compose.material.MaterialTheme
-import androidx.compose.material.ProvideEmphasis
 import androidx.compose.material.Surface
+import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Bookmark
 import androidx.compose.material.icons.filled.BookmarkBorder
 import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Providers
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.imageResource
@@ -52,7 +52,7 @@ fun AuthorAndReadTime(
     modifier: Modifier = Modifier
 ) {
     Row(modifier) {
-        ProvideEmphasis(AmbientEmphasisLevels.current.medium) {
+        Providers(AmbientContentAlpha provides ContentAlpha.medium) {
             val textStyle = MaterialTheme.typography.body2
             Text(
                 text = post.metadata.author.name,
@@ -79,9 +79,7 @@ fun PostImage(post: Post, modifier: Modifier = Modifier) {
 
 @Composable
 fun PostTitle(post: Post) {
-    ProvideEmphasis(AmbientEmphasisLevels.current.high) {
-        Text(post.title, style = MaterialTheme.typography.subtitle1)
-    }
+    Text(post.title, style = MaterialTheme.typography.subtitle1)
 }
 
 @Composable
@@ -118,7 +116,7 @@ fun PostCardHistory(post: Post, navigateTo: (Screen) -> Unit) {
             modifier = Modifier.padding(end = 16.dp)
         )
         Column(Modifier.weight(1f)) {
-            ProvideEmphasis(AmbientEmphasisLevels.current.medium) {
+            Providers(AmbientContentAlpha provides ContentAlpha.medium) {
                 Text(
                     text = "BASED ON YOUR HISTORY",
                     style = MaterialTheme.typography.overline
@@ -130,8 +128,8 @@ fun PostCardHistory(post: Post, navigateTo: (Screen) -> Unit) {
                 modifier = Modifier.padding(top = 4.dp)
             )
         }
-        ProvideEmphasis(AmbientEmphasisLevels.current.medium) {
-            Icon(asset = Icons.Filled.MoreVert)
+        Providers(AmbientContentAlpha provides ContentAlpha.medium) {
+            Icon(Icons.Filled.MoreVert)
         }
     }
 }
@@ -144,19 +142,13 @@ fun BookmarkButton(
 ) {
     IconToggleButton(
         checked = isBookmarked,
-        onCheckedChange = { onClick() }
+        onCheckedChange = { onClick() },
+        modifier = modifier
     ) {
-        modifier.fillMaxSize()
         if (isBookmarked) {
-            Icon(
-                asset = Icons.Filled.Bookmark,
-                modifier = modifier
-            )
+            Icon(asset = Icons.Filled.Bookmark)
         } else {
-            Icon(
-                asset = Icons.Filled.BookmarkBorder,
-                modifier = modifier
-            )
+            Icon(asset = Icons.Filled.BookmarkBorder)
         }
     }
 }